+++ tests/dummy/app/controllers/components-examples/flexberry-objectlistview/list-on-editform.js
@@ -1,0 +1,24 @@
+import ListFormController from 'ember-flexberry/controllers/list-form';
+
+export default ListFormController.extend({
+  /**
+    Name of related edit form route.
+
+    @property editFormRoute
+    @type String
+    @default 'ember-flexberry-dummy-suggestion-edit'
+   */
+  editFormRoute: 'components-examples/flexberry-objectlistview/on-edit-form/suggestion',
+
+  exportExcelProjection: 'SuggestionL',
+
+  actions: {
+    /**
+      Hook that executes before deleting all records on all pages.
+      Need to be overriden in corresponding application controller.
+    */
+    beforeDeleteAllRecords(modelName, data) {
+      data.cancel = false;
+    }
+  }
+});