+++ tests/dummy/app/locales/ru/translations.js
@@ -452,6 +452,10 @@ Ember.$.extend(true, translations, {
               'caption': 'Размещение на форме редактирования',
               'title': ''
             },
+            'list-on-editform': {
+              'caption': 'Размещение списка детейлов мастера на форме редактирования',
+              'title': ''
+            },
             'custom-filter': {
               'caption': 'Настройка фильтра',
               'title': ''
@@ -883,6 +887,9 @@ Ember.$.extend(true, translations, {
         'on-edit-form': {
           'caption': 'Flexberry-objectlistview. FlexberryObjectlistview на форме редактирования',
           'add-button-name': 'Добавить'
+        },
+        'list-on-editform': {
+          'caption': 'Список детейлов Тип предложения'
         },
         'custom-filter': {
           'caption': 'Flexberry-objectlistview. Настройка фильтра'