+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.12.3] - 2018-09-18
+### Fixed
+* `flexberry-lookup` component:
+    * Fix list's hierarchy when several lookups on page.
+* `list-form` route:
+    * Fix transition to list-form with custom query params.
+    
+
 ## [0.12.2] - 2018-08-28
 ### Fixed
 * Missing of calling `_super` method in ember hooks inside addon.